Bug warnings fixes (#470)
* host change

* fixed shadow warning

* made wantedView mobileView a different component

* made wantedView desktopView a component of its own

* added mobile view for wantedSummary

* added Buttons to component file

* created seperate components for bounty tasks

* added codingDesktop

* fixed merge conflicts

* fixed merge conflicts

* fixed merged conflicts and added eslint-disable func-style

* made style props small letters

* restored MODE positon

* fixed more warnings

* fixed edit bounty bugs

* changed hosts to default

* fixed merge conflicts

* fixed merge conflicts

* fixed merge conflicts

* fixed merge conflicts

* moved validator code to utils.ts

* moved validator code to utils.ts

* placed personViewSlim styles in a new file

* made person slim mobile file

* added MobileView to personView

* DesktopView added to person slim

* removed multiple console.log statements

* changed to defaut App.tsx

---------
