+++ frontend/app/src/api/index.ts
@@ -51,7 +51,6 @@ function addMethod(m: string): Function {
       if (m === 'BLOB') opts.method = 'GET';
       const r = await fetch(rootUrl + url, opts);
       if (!r.ok) {
-        // console.log(r)
         throw new Error('Not OK!');
       }
       let res;
