+++ frontend/app/src/components/form/inputs/img-input.tsx
@@ -30,7 +30,7 @@ export default function ImageInput({
       const info = ui.meInfo as any;
       if (!info) {
         alert('You are not logged in.');
-        return console.log('no meInfo');
+        return;
       }
       const URL = info.url.startsWith('http') ? info.url : `https://${info.url}`;
       const r = await fetch(`${URL}/public_pic`, {
@@ -56,7 +56,6 @@ export default function ImageInput({
   }
 
   async function dropzoneUpload(files: File[], fileRejections) {
-    console.log('fileRejections', fileRejections);
     if (fileRejections.length) {
       fileRejections.forEach((file) => {
         file.errors.forEach((err) => {
@@ -72,7 +71,6 @@ export default function ImageInput({
       return;
     }
 
-    console.log(files);
     const file = files[0];
     setUploading(true);
     const reader = new FileReader();
@@ -85,8 +83,8 @@ export default function ImageInput({
 
   const addedStyle = notProfilePic
     ? {
-        borderRadius: 0
-      }
+      borderRadius: 0
+    }
     : {};
 
   const defaultIcon = notProfilePic ? backgroundIcon : avatarIcon;
@@ -103,9 +101,8 @@ export default function ImageInput({
                   {!uploading ? (
                     <Image
                       style={{
-                        backgroundImage: `url(${
-                          picsrc ? picsrc : value ? value : uploading ? '' : defaultIcon
-                        })`,
+                        backgroundImage: `url(${picsrc ? picsrc : value ? value : uploading ? '' : defaultIcon
+                          })`,
                         ...addedStyle
                       }}
                     />
@@ -143,9 +140,8 @@ export default function ImageInput({
                   {!uploading ? (
                     <Image
                       style={{
-                        backgroundImage: `url(${
-                          picsrc ? picsrc : value ? value : uploading ? '' : defaultIcon
-                        })`,
+                        backgroundImage: `url(${picsrc ? picsrc : value ? value : uploading ? '' : defaultIcon
+                          })`,
                         ...addedStyle
                       }}
                     />
