+++ frontend/app/src/people/auth/authQR.tsx
@@ -33,7 +33,6 @@ export default function AuthQR(props: any) {
     interval = setInterval(async () => {
       try {
         const me: MeInfo = await api.get(`poll/${challenge}`);
-        console.log(me);
         if (me && me?.pubkey) {
           ui.setMeInfo(me);
           await main.getSelf(me);
@@ -45,7 +44,7 @@ export default function AuthQR(props: any) {
         if (i > 100) {
           if (interval) clearInterval(interval);
         }
-      } catch (e) {}
+      } catch (e) { }
     }, 3000);
   }
   async function getChallenge() {
