+++ frontend/app/src/people/main/body.tsx
@@ -123,7 +123,6 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
   const publicPanelClick = useCallback(
     async (person, item) => {
       // migrating to load widgets separate from person
-      console.log('person', { person }, 'and items', { item });
       const itemIndex = person[selectedWidget]?.findIndex((f) => f.created === item.created);
       if (itemIndex > -1) {
         // make person into proper structure (derived from widget)
@@ -149,16 +148,16 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
       const value =
         activeList && activeList.length
           ? activeList.find((item) => {
-              const { person, body } = item;
-              return owner_id === person.owner_pubkey && created === `${body.created}`;
-            })
+            const { person, body } = item;
+            return owner_id === person.owner_pubkey && created === `${body.created}`;
+          })
           : {};
       setActiveListIndex(
         activeList && activeList.length
           ? activeList.findIndex((item) => {
-              const { person, body } = item;
-              return owner_id === person.owner_pubkey && created === `${body.created}`;
-            })
+            const { person, body } = item;
+            return owner_id === person.owner_pubkey && created === `${body.created}`;
+          })
           : {}
       );
 
@@ -231,7 +230,6 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
   };
 
   function selectPerson(id: number, unique_name: string, pubkey: string) {
-    console.log('selectPerson', id, unique_name, pubkey);
     ui.setSelectedPerson(id);
     ui.setSelectingPerson(id);
 
@@ -441,7 +439,6 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
                     background: color.pureWhite
                   }}
                   onChange={(e) => {
-                    console.log('handleChange', e);
                     ui.setSearchText(e);
                   }}
                 />
@@ -505,8 +502,8 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
     const focusedDesktopModalStyles =
       selectedWidget && widgetConfigs[selectedWidget]
         ? {
-            ...widgetConfigs[selectedWidget].modalStyle
-          }
+          ...widgetConfigs[selectedWidget].modalStyle
+        }
         : {};
 
     // desktop mode
@@ -557,7 +554,6 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
                 background: color.grayish.G600
               }}
               onChange={(e) => {
-                console.log('handleChange', e);
                 ui.setSearchText(e);
               }}
             />
@@ -634,7 +630,7 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
               activeListIndex === 0
                 ? null
                 : isModalSideButton
-                ? () => {
+                  ? () => {
                     const { person, body } = activeList[activeListIndex - 1];
                     if (person && body) {
                       history.replace({
@@ -647,13 +643,13 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
                       });
                     }
                   }
-                : null
+                  : null
             }
             nextArrowNew={
               activeListIndex + 1 > activeList?.length
                 ? null
                 : isModalSideButton
-                ? () => {
+                  ? () => {
                     const { person, body } = activeList[activeListIndex + 1];
                     if (person && body) {
                       history.replace({
@@ -666,7 +662,7 @@ export default function BodyComponent({ selectedWidget }: { selectedWidget: Widg
                       });
                     }
                   }
-                : null
+                  : null
             }
           >
             <FocusedView
