+++ frontend/app/src/tribes/body.tsx
@@ -50,8 +50,6 @@ export default function BodyComponent() {
     let newPage = currentPage + direction;
     if (newPage < 1) newPage = 1;
 
-    console.log('loadmore');
-
     try {
       await main.getTribes({ page: newPage });
     } catch (e) {
@@ -61,7 +59,7 @@ export default function BodyComponent() {
 
   async function refreshList() {
     setLoadingList(true);
-    console.log('refreshList');
+
     // reset page will replace all results, this is good for a new search!
     await main.getTribes({ page: 1, resetPage: true });
 
@@ -197,7 +195,6 @@ export default function BodyComponent() {
                 marginLeft: 20
               }}
               onChange={(e) => {
-                console.log('handleChange', e);
                 ui.setSearchText(e);
               }}
             />
